blog update, delete completed

+++ src/services/blogCategory/DeleteBlogCategoryService.ts
@@ -1,5 +1,6 @@
 import CustomError from "../../errors/CustomError";
 import BlogCategoryModel from "../../models/BlogCategoryModel";
+import BlogModel from "../../models/BlogModel";
 import isNotObjectId from "../../utils/isNotObjectId";
 
 const DeleteBlogCategoryService = async (categoryId: string) => {
@@ -11,13 +12,13 @@ const DeleteBlogCategoryService = async (categoryId: string) => {
         throw new CustomError(404, 'This categoryId not found');
     }
 
-    //check if categoryId is associated with subCategory
-    // const associatedWithSubCategory = await SubCategoryModel.findOne({
-    //      categoryId
-    // });
-    // if(associatedWithSubCategory){
-    //     throw new CustomError(409, 'Unable to delete, This category is associated with sub-category.');
-    // }
+    //check if categoryId is associated with blog
+    const associatedWithBlog = await BlogModel.findOne({
+         categoryId
+    });
+    if(associatedWithBlog){
+        throw new CustomError(409, 'Unable to delete, This category is associated with blog.');
+    }
 
     const result = await BlogCategoryModel.deleteOne({ _id: categoryId})
     return result;
